feature: Add temporary directory creation on Dir
TL;DR - This is a monkey-patch to create temporary directories with
Crystal standard library.

The PR is open, but still solving some issues.
https://github.com/crystal-lang/crystal/pull/2911

The mkdtemp function creates a directory with a unique name. If it
succeeds, it overwrites template with the name of the directory, and
returns template. As with mktemp and mkstemp, template should be a
string ending with ‘XXXXXX’.

The directory is created using mode 0700

References:
  - http://www.gnu.org/software/libc/manual/html_node/Temporary-Files.html
  - http://ruby-doc.org/stdlib-2.0.0/libdoc/tmpdir/rdoc/Dir.html